+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.201 ] - 2025/04/26
+
+  * Added logic to resolve device object by device name when there are duplicate names defined to the device list.  If the resolved device name is the same as the active device name, but the id's are different, then the active device is used instead of the resolved device.
+  * Updated `SpotifyConnectDirectoryTask.ActivateCastAppSpotify` to search for device by id first; if no matches, then search for device by name.
+
 ###### [ 1.0.200 ] - 2025/04/23
 
   * Added logic to detect device name changes when speakers are grouped / ungrouped. This can happen for devices that are added to / removed from groups and do not correctly inform interested parties via a zeroconf OnServiceStateChange event that the name has changed (e.g. Denon HEOS devices, etc).